requirements.txt
numpy
ftfy
regex
tqdm
omegaconf
pytorch-lightning  
kornia
einops
wget
stegano               
python-xmp-toolkit    
imgtag                
pillow        
imageio-ffmpeg
matplotlib
scikit-learn
openai
scikit-image
torch
torchvision
ipython
librosa
moviepy
jupyter 
ipywidgets
transformers